backend/requirements.txt
fastapi
uvicorn
python-dotenv
supabase
pydantic
email-validator
requests
greenery
google-generativeai
